requirements.txt
aiomysql==0.0.21
discord.py[voice]==1.6.0
PyYAML==5.4.1
sqlalchemy==1.4.4